core: added support to skip paging in data table

+++ src/Montr.Core/Models/SearchResult.cs
@@ -4,7 +4,7 @@ namespace Montr.Core.Models
 {
 	public class SearchResult<TModel> // : ApiResult (?)
 	{
-		public int? TotalCount { get; set; }
+		public long? TotalCount { get; set; }
 
 		public IList<TModel> Rows { get; set; }
 	}
